test(host-infra-system): split git adapter tests by concern
Move git tests into focused modules with shared support helpers, colocate parser/helper unit tests with their source modules, and add parser edge-case coverage for diff header and ahead/behind parsing.
