+++ apps/desktop/src-tauri/crates/host-infra-system/src/git/tests/mod.rs
@@ -1,0 +1,7 @@
+mod branch_tests;
+mod commit_tests;
+mod remote_tests;
+mod status_tests;
+mod support;
+mod validation_tests;
+mod worktree_tests;
